requirements.txt
Flask==2.3.3
pathway[xpack-llm]==0.7.6
python-dotenv==1.0.0
pandas==2.1.1
numpy==1.24.3
requests==2.31.0
openai==0.28.1
cohere==4.32
transformers==4.34.1
torch==2.0.1
accelerate==0.23.0
litellm==1.0.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3